requirements.txt
asgiref==3.5.2
black==22.3.0
click==8.1.3
Django==4.0.5
django-ckeditor==6.4.2
django-debug-toolbar==3.5.0
django-js-asset==2.0.0
flake8==4.0.1
isort==5.10.1
lark-parser==0.12.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-decouple==3.6
required==0.4.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
whitenoise==6.2.0
